Added functions to evaluate the reached audience together with the corresponding cost of each jsonb array.
weighted_audience multiplies the audience with the weighted cost.
compute_audience_index just sums up all weighted audience values.

As of now the index represents the number of reached audience from within one cell. BUT there has been weights assigned, so for example, for a walking time of 6 minutes, only half of the audience is considered in the index.